refactoring. add cloning repositories

+++ init/windows.ps1
@@ -24,39 +24,63 @@ $chocolatey_packages = @(
     "microsoft-windows-terminal"
 )
 
+$manual_build_list = @(
+    @{"name" = "vim"; "url" = "https://github.com/vim/vim"},
+    @{"name" = "juce"; "url" = "https://github.com/juce-framework/juce"},
+    @{"name" = "frut"; "url" = "https://github.com/mcmartin/frut"}
+)
 
 function install_scoop_packages {
-    echo "install scoop packages"
-    if (-Not(gcm scoop -ea SilentlyContinue)){
-        Invoke-Expression (New-Object System.Net.WebClient).DownloadString('https://get.scoop.sh')
+    Write-Output "install scoop packages"
+    if (-not(Get-Command scoop -ea silentlycontinue)){
+        Invoke-Expression (New-Object system.net.webclient).downloadstring('https://get.scoop.sh')
     }
     foreach( $package in $scoop_packages ){
-        if (scoop info $package | sls "Installed: No"){
-            echo "${package} is not installed. install ${package}"
+        if (scoop info $package | Select-String "installed: no"){
+            Write-Output "${package} is not installed. install ${package}"
             scoop install $package
         }else{
-            echo "$package is installed or can't find manifest. skip"
+            Write-Output "$package is installed or can't find manifest. skip"
         }
     }
 }
 
 function install_choco_packages {
-    echo "install chocolatey packages"
-    if (-Not(gcm choco -ea SilentlyContinue)){
-        Invoke-Expression (New-Object System.Net.WebClient).DownloadString('https://get.scoop.sh')
+    Write-Output "install chocolatey packages"
+    if (-not(Get-Command choco -ea silentlycontinue)){
+        Invoke-Expression (New-Object system.net.webclient).downloadstring('https://get.scoop.sh')
     }
     foreach( $package in $chocolatey_packages){
-        if (-Not(choco list -l $package | sls  "(chocolatey|packages installed)" -NotMatch)){
-            echo "${package} is not installed. install ${package}"
+        if (-not(choco list -l $package | Select-String  "(chocolatey|packages installed)" -notmatch)){
+            Write-Output "${package} is not installed. install ${package}"
             cinst -y $package
         }else{
-            echo "$package is installed. skip"
+            Write-Output "$package is installed. skip"
         }
+    }
+}
+
+function clone_manual_build_repository {
+    if (-Not(Get-Command git -ea SilentlyContinue)){
+        return
+    }
+    if (-Not(Test-Path ".\manual_build_repositories")){
+        New-Item -ItemType Directory manual_build_repositories
+    }
+    foreach( $repo in $manual_build_list ){
+        $clone_path = (Join-Path ".\manual_build_repositories" $repo.name)
+        $repo_name = $repo.name
+        if (Test-Path $clone_path){
+            Write-Output "$repo_name has already cloned. skip"
+            continue
+        }
+        git clone $repo.url $clone_path
     }
 }
 
 install_scoop_packages
 install_choco_packages
+clone_manual_build_repository
 
 $default_apps = @(
  'Microsoft.Microsoft3DViewer',
@@ -97,6 +121,9 @@ $default_apps = @(
 )
 
 function remove_default_applications {
+    if (-Not(Get-Command Get-AppxPackage -ea SilentlyContinue)){
+        return
+    }
     foreach( $app in $default_apps ){
         Get-AppxPackage $app | Remove-AppxPackage
     }
